Add zephyrine.Positionable typeclass; unify locate/locateKey
`locate`/`locateKey` (resolve a path to a source Position in tracked data) were
duplicated as top-level extensions in both jacinta (on Json/JsonPointer) and
ypsiloid (on Yaml/YamlPath). Duplicate top-level names can't both be re-exported
into the `soundness` umbrella (they'd clash), and SN-742.1 flagged them.

Introduce a `Positionable` typeclass in zephyrine that links a data type to its
path type (`by`, the Operand) and its position type (`to`, the Result):
`Json is Positionable by JsonPointer to Json.Ast.Position`. A single pair of
`value.locate(path)`/`value.locateKey(path)` extensions in zephyrine_core
dispatches through it, exported once. jacinta and ypsiloid now provide instances
(in their `object Json`/`object Yaml` companions, alongside the moved walkIndex
helper) instead of top-level extensions, and their `*_position.scala` files are
removed. Other tracking-aware formats can join by supplying an instance.
